Change int to size_t in HttpResponseParser
r=bmcquade

+++ src/mod_spdy/common/http_response_parser.h
@@ -73,7 +73,7 @@ class HttpResponseParser {
   HttpResponseVisitorInterface* const visitor_;
   ParserState state_;
   BodyType body_type_;
-  int remaining_bytes_;
+  size_t remaining_bytes_;
   std::string buffer_;
 
   DISALLOW_COPY_AND_ASSIGN(HttpResponseParser);
